S0C-3A-1A: Add shadow_verify_dual_run_stage2 scenario and update registry
- Introduced a new scenario `shadow_verify_dual_run_stage2` to enhance dual run capabilities.
- Updated the registry to include the new scenario for loading built-in scenarios.
- Documented the results of the readiness gate run in the logs, confirming successful transitions and validations.
