+++ docs/logs/log-S0C-3A-1A-double-parallel.md
@@ -159,11 +159,22 @@
   - `es_health_ok=true`, `backfill_ok=true`, `es_search_ok=true`
   - `parity_ok=true`
 
+### Run 6 — readiness_gate（OK, shim→scenario 后复跑）
+
+- artifact: `summary.json`
+- scenario: `shadow_verify_dual_run_readiness_gate`
+- run_id: `22255455516-1`
+- result: `ok=true`
+- 关键输出（节选）:
+  - `checks.write_gate.ok=true`（引用子检查结果路径）
+  - `checks.paging_stability.ok=true`
+  - `checks.shared_keys.ok=true`
+
 ## Status Update（阶段结论）
 
 - Step B（迁移优先级前三个场景 + readiness gate 复合验证）已通过实证：`Run 1~4` 均 `ok=true`。
 - dual-run（stage1）已迁移并在 GitHub Actions 中跑通：`Run 5` 为 `ok=true`（strict 对齐）。
-- readiness gate 已具备 “shim → scenario” 转接基础，可继续按优先级推进 stage2/window/canary 等迁移，同时保持对外契约不变。
+- readiness gate 已完成 “shim → scenario” 转接并复跑通过：`Run 6` 为 `ok=true`。
 
 
 ## Risks（风险与缓解）
